feat(canvas): canvas抽取两个元应用
