+++ packages/canvas/breadscrumb/index.js
@@ -1,0 +1,6 @@
+import CanvasBreadcrumb from './src/CanvasFooter.vue'
+import metaData from './meta'
+export default {
+  metaData,
+  component: CanvasBreadcrumb
+}
