+++ packages/canvas/breadscrumb/meta.js
@@ -1,0 +1,3 @@
+export default {
+  id: 'engine.canvas.breadcrumb'
+}
