+++ packages/canvas/canvas.html
@@ -207,6 +207,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/canvas.js"></script>
+    <script type="module" src="./test/src/canvas.js"></script>
   </body>
 </html>
