+++ packages/canvas/common/src/constant.js
@@ -1,0 +1,3 @@
+export const NODE_UID = 'data-uid'
+export const NODE_TAG = 'data-tag'
+export const NODE_LOOP = 'loop-id'
