+++ packages/canvas/container/index.js
@@ -1,0 +1,8 @@
+import CanvasContainer from './src/CanvasContainer.vue'
+import metaData from './meta'
+import api from './api'
+export default {
+  ...metaData,
+  component: CanvasContainer,
+  api
+}
