+++ packages/canvas/container/src/api/index.js
@@ -1,0 +1,2 @@
+import api from './useCanvas'
+export default api
