+++ packages/canvas/container/src/components/CanvasDivider.vue
