+++ packages/canvas/container/src/components/CanvasResize.vue
