+++ packages/canvas/container/src/components/CanvasResizeBorder.vue
