+++ packages/canvas/drag-item/index.js
@@ -1,0 +1,2 @@
+import CanvasDragItem from './src/CanvasDragItem.vue'
+export default CanvasDragItem
