+++ packages/canvas/drag-item/src/CanvasDragItem.vue
