+++ packages/canvas/index.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/main.js"></script>
+    <script type="module" src="./test/src/main.js"></script>
   </body>
 </html>
