+++ packages/canvas/render/index.js
@@ -1,0 +1,1 @@
+export { createRender } from './src/runner'
