+++ packages/canvas/render/src/CanvasEmpty.vue
