+++ packages/canvas/render/src/RenderMain.js
