+++ packages/canvas/render/src/builtin/CanvasBox.vue
