+++ packages/canvas/render/src/builtin/CanvasCollection.vue
@@ -14,7 +14,7 @@ import CanvasPlaceholder from './CanvasPlaceholder.vue'
 import { getHandler } from './CanvasCollection.js'
 
 export const fetchDataSourceDetail = (dataSourceId) =>
-  getController().request.get(`/app-center/api/sources/detail/${dataSourceId}`)
+  getController().request.get(`/app-center/api/sources/detail/${dataSourceId}`) // TODO: 强行耦合了
 
 export default {
   components: {
