+++ packages/canvas/render/src/builtin/CanvasImg.vue
