+++ packages/canvas/render/src/builtin/CanvasPlaceholder.vue
