+++ packages/canvas/render/src/builtin/CanvasText.vue
