+++ packages/canvas/test/src/i18n/en.json
