+++ packages/canvas/test/src/i18n/zh.json
