+++ packages/canvas/test/src/main.js
@@ -11,24 +11,7 @@
  */
 
 import { createApp } from 'vue'
-import * as TinyVue from '@opentiny/vue'
 
-import { RenderMain } from './index'
+import App from './Design.vue'
 
-import { I18nInjectionKey, createI18n } from 'vue-i18n'
-
-window.TinyLowcodeComponent = {}
-
-Object.entries(TinyVue).forEach(([_key, component]) => {
-  const { name } = component
-  if (name) {
-    window.TinyLowcodeComponent[name] = component
-  }
-})
-
-const i18nHost = createI18n({
-  locale: 'zh_CN',
-  messages: {}
-})
-
-createApp(RenderMain).use(i18nHost).provide(I18nInjectionKey, i18nHost).mount('#app')
+createApp(App).mount('#app')
